+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Up Next
 Dockerfiles. This is necessary to fetch updates when Dockerfiles are
 non-deterministic and rely on pulling data from the network.
 - Use the latest stable release of Docker Engine.
+- Fixed a bug where `quilt inspect` would panic when given a relative path.
 
 Release 0.2.0
 -------------